split up parentI3 so that it can be reused
