+++ tests/unit/libpluto/lp13-parentI3/Makefile.testcase
@@ -1,0 +1,6 @@
+TESTNAME=parentI3
+CONNNAME=parker1--jj2
+WHACKFILE=${OUTPUTS}/ikev2client.record.${ARCH}
+UNITTEST1ARGS=-r ${WHACKFILE} ${CONNNAME} ../lp10-parentI2/parentR1.pcap parentR2.pcap
+ENDNAME=parker
+
